aragwas_server/requirements.txt
appdirs==1.4.3
Django==1.11b1
djangorestframework==3.6.2
packaging==16.8
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
djangorestframework-csv==2.0.0
h5py==2.6.0
celery==4.0.2
coreapi==2.3.0
pytest-django==3.1.2
elasticsearch-dsl>=5.0.0,<6.0.0
-e git+git@github.com:timeu/gff3-py.git#egg=gff3
